Fixed small comment.
A note to you if you're still awake: I had almost the full win logic completed and my laptop died. Usually it saves my progress locally, but it didn't this time for whatever reason and I was in the middle of pushing to github, so my file corrupted and everything I wrote over the last 3 hours is just gone. I pulled your file from the github, but my progress on the win/lose stuff is now gone. I'll rewrite it and push it tonight, I am unbelievably pissed at it for doing that.