Final changes made!
Added introductory text to sidebar
Formatted entire sidebar
Added final Javadoc comment changes

I am now calling this officially done for the time being. Everything is implemented that we need to have, all Javadoc comments are written, code is formatted and readable. Many thanks for the work on the game logic and internal side of things.